+++ src/circuit/gadgets/mod.rs
@@ -2,9 +2,11 @@
 //!
 //! This module contains low-level circuit components used in the synthesis.
 
-pub mod hash;
 pub mod merkle;
+pub mod poseidon;
+pub mod select;
 
 // Re-export commonly used gadgets
-pub use hash::{conditional_select, poseidon_hash_tagged_gadget};
 pub use merkle::{verify_aggregation_path_gated, verify_merkle_path_gated};
+pub use poseidon::poseidon_hash_tagged_gadget;
+pub use select::conditional_select;
